pyproject.toml
[project]
name = "mcp-server-opendal"
version = "0.1.0"
description = "A Model Context Protocol server providing tools to access storage services for usage by LLMs"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
 "mcp>=1.0.0",
 "opendal>=0.45.16,<0.46.0",
 "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["ruff>=0.7.3"]

[project.scripts]
mcp-server-opendal = "mcp_server_opendal:main"
